+++ instrumentation/github.com/labstack/echo/otelecho/echo_test.go
@@ -117,3 +117,25 @@ func TestSkipper(t *testing.T) {
 	router.ServeHTTP(w, r)
 	assert.Equal(t, http.StatusOK, w.Result().StatusCode, "should call the 'ping' handler") //nolint:bodyclose // False positive for httptest.ResponseRecorder: https://github.com/timakin/bodyclose/issues/59.
 }
+
+func TestHandleErrorHandlerNotCalled(t *testing.T) {
+	r := httptest.NewRequest("GET", "/ping", nil)
+	w := httptest.NewRecorder()
+
+	router := echo.New()
+	router.Use(Middleware("foobar"))
+	router.GET("/ping", func(c echo.Context) error {
+		return assert.AnError
+	})
+
+	handlerCalled := 0
+	router.HTTPErrorHandler = func(err error, c echo.Context) {
+		handlerCalled++
+		assert.ErrorIs(t, err, assert.AnError, "test error is expected in error handler")
+		assert.NoError(t, c.NoContent(http.StatusTeapot))
+	}
+
+	router.ServeHTTP(w, r)
+	assert.Equal(t, http.StatusTeapot, w.Result().StatusCode, "should call the 'ping' handler")
+	assert.Equal(t, 1, handlerCalled, "global error handler must be called only once")
+}